Improve indexing of NXlog
